# openapi.yaml
openapi: 3.0.0
servers:
  - url: https://api.kalastud.io
    description: Production Server
  - url: https://perfect-week-test.herokuapp.com
    description: Test Server
info:
  title: PerfectWeek API
  description: The PerfectWeek web API
  version: 0.1.0
  contact:
    email: perfectweek_2020@labeip.epitech.eu
  license:
    name: MIT
    url: 'https://github.com/PerfectWeek/web-api/blob/dev/LICENSE'

####################
# Tags
#
...
paths:
...
  /events/{event-id}:
    get:
      tags:
        - Event
      summary: Get information about an Event
      operationId: getEventInfo
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: event-id
          required: true
          description: The Event id
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: OK
                  event:
                    $ref: '#/components/schemas/EventInfo'
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationFailed'
        '403':
          description: Action not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: Event not Found
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'

    put:
      tags:
        - Event
      summary: Edit an Event information
      operationId: editEventInfo
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/EventEdit'
      parameters:
        - in: path
          name: event-id
          required: true
          description: The Event id
          schema:
            type: string
      responses:
        '200':
          description: Event successfully edited
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: OK
                  event:
                    $ref: '#/components/schemas/EventInfo'
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationFailed'
        '403':
          description: Action not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'

    delete:
      tags:
        - Event
      summary: Delete an Event
      operationId: deleteEvent
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: event-id
          required: true
          description: The Event id
          schema:
            type: string
      responses:
        '200':
          description: Event successfully deleted
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: OK
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationFailed'
        '403':
          description: Action not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: Event not found
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
...
  /events/{event-id}/upload-image:
    post:
      tags:
        - Event
      summary: Upload image as new Avatar for specific event
      operationId: uploadEventImage
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: event-id
          required: true
          description: The Event id
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
            encoding:
              image:
                contentType: image/png, image/jpeg
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: OK
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationFailed'
        '403':
          description: Action not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
        '413':
          description: Entity Too Large
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: Image should not exceed 2MB
  /events/{event-id}/image:
    get:
      tags:
        - Event
      summary: Get event profile image
      operationId: getEventImage
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: event-id
          required: true
          description: The Event id
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: OK
                  image:
                    type: string
                    example: 'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAfQAAAH0CAIAAABEtEjdAAABG2l...'
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationFailed'
        '403':
          description: Action not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
 #
  # Search
  #
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  #
  # Schemas
  #
  schemas:
...
    EventInfo:
      type: object
      properties:
        id:
          type: number
          example: 43
        name:
          type: string
          example: nouvel an
        description:
          type: string
          example: soiree
        location:
          type: string
          example: Le pub Universitaire, Quebec City
        type:
          type: string
          example: party
        visibility:
          type: string
          example: public
        calendar_id:
          type: number
          example : 3
        start_time:
          type: string
          example: "2012-12-12T12:12:12"
        end_time:
          type: string
          example: "2012-12-12T12:12:12"
        status:
          type: string
          example: invited
    EventEdit:
      type: object
      properties:
        name:
          type: string
          example: nouvel an
        description:
          type: string
          example: soiree
        type:
          type: string
          example: party
        visibility:
          type: string
          example: public
        start_time:
          type: string
          example: "2012-12-12T12:12:12"
        end_time:
          type: string
          example: "2012-12-12T12:12:12"
        location:
          type: string
          example: "Bastille"
...
    AuthenticationFailed:
      description: The authentication failed
      type: object
      properties:
        message:
          type: string
          example: You must be authenticated to perform this action
    # 403
    Unauthorized:
      description: The action is not allowed
      type: object
      properties:
        message:
          type: string
          example: Action not allowed
    # 404
...
    InternalServerError:
      description: Internal Server Error
      type: object
      properties:
        message:
          type: string
          example: Internal Server Error